Moved tests to respective tests classes
